+++ app/Http/Controllers/RolePermissionController.php
@@ -10,6 +10,7 @@ use App\Models\Permission;
 use App\Models\Role;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Str;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
@@ -18,6 +19,7 @@ class RolePermissionController extends Controller
 {
     public function connect()
     {
+        Gate::authorize('has-role',UserRole::SUPER_ADMIN);
         $users = User::whereNotIn('id', [1, 2])->get();
         // Create the permission if it doesn't exist
         $permissionName = UserPermission::SCRIPTORIUM_PERMISSIONS->value;
@@ -32,6 +34,7 @@ class RolePermissionController extends Controller
 
     public function table(Request $request)
     {
+        Gate::authorize('has-role',UserRole::SUPER_ADMIN);
         $user = Auth::user();
 
         // Role query builder
@@ -74,6 +77,7 @@ class RolePermissionController extends Controller
     }
     public function create_role()
     {
+        Gate::authorize('has-role',UserRole::SUPER_ADMIN);
         $permissions = DB::table('permissions')->select('id','name')->paginate(5);
         return view('permission.create-role',[
             'permissions' => $permissions
@@ -81,6 +85,7 @@ class RolePermissionController extends Controller
     }
     public function store_role(RoleStoreRequest $request)
     {
+        Gate::authorize('has-role',UserRole::SUPER_ADMIN);
         $request->validated();
 
         $role = Role::create(['name' => $request->role]);
@@ -91,12 +96,14 @@ class RolePermissionController extends Controller
     }
     public function edit_role(string $id)
     {
+        Gate::authorize('has-role',UserRole::SUPER_ADMIN);
         $role = Role::findOrFail($id);
         $permissions = Permission::select('id', 'name')->get();
         return view('permission.edit-role',['role'=>$role , 'permissions'=> $permissions]);
     }
     public function update_role(RoleStoreRequest $request, string $id)
     {
+        Gate::authorize('has-role',UserRole::SUPER_ADMIN);
         $request->validated();
         $role = Role::findOrFail($id);
         $role->name = $request->role;
@@ -106,6 +113,7 @@ class RolePermissionController extends Controller
     }
     public function destroy_role(string $id)
     {
+        Gate::authorize('has-role',UserRole::SUPER_ADMIN);
         $role = Role::findOrFail($id)->delete();
         DB::table('permission_role')->where('role_id',$id)->delete();
         return to_route('role.permission.table')->with('status','نقش با موفقیت حذف شد');
@@ -114,27 +122,27 @@ class RolePermissionController extends Controller
     // these below are Permission
     public function create_permission()
     {
+        Gate::authorize('has-role',UserRole::SUPER_ADMIN);
         return view('permission.create-permission');
     }
     public function store_permission(PermissionStoreRequest $request)
     {
+        Gate::authorize('has-role',UserRole::SUPER_ADMIN);
         $request->validated();
         $permission = Str::squish($request->permission);
         Permission::create(['name' => $permission]);
         return to_route('role.permission.table')->with('status','سطح دسترسی جدید با موفقیت ایجاد شد');
     }
-//    public function show_permission(string $id)
-//    {
-//        $permission = Permission::findOrFail($id);
-//        return view('permission.show-permission',['permission'=>$permission]);
-//    }
+
     public function edit_permission(string $id)
     {
+        Gate::authorize('has-role',UserRole::SUPER_ADMIN);
         $permission = Permission::findOrFail($id);
         return view('permission.edit-permission',['permission'=> $permission]);
     }
     public function update_permission(PermissionStoreRequest $request, string $id)
     {
+        Gate::authorize('has-role',UserRole::SUPER_ADMIN);
         $request->validated();
         $permissionName = Str::squish($request->permission);
         $permission = Permission::findOrFail($id);
@@ -144,10 +152,9 @@ class RolePermissionController extends Controller
     }
     public function destroy_permission(string $id)
     {
+        Gate::authorize('has-role',UserRole::SUPER_ADMIN);
         $permission = Permission::findOrFail($id)->delete();
         return to_route('role.permission.table')->with('status','سطح دسترسی با موفقیت حذف شد');
     }
-
-
 
 }
